lab-1/requirements.txt
# Constraints for Airflow 3.0.6 with Python 3.12
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-3.0.6/constraints-3.12.txt"

# DAG Factory for declarative DAG creation from YAML
dag-factory==1.0.0

# FAB Auth Manager (required for user authentication in Airflow 3.0+)
apache-airflow-providers-fab>=1.0.0

# Additional useful packages for Airflow development
pyyaml>=6.0
